+++ exonum-java-binding-cryptocurrency-demo/src/main/java/com/exonum/binding/cryptocurrency/CryptocurrencyService.java
@@ -16,10 +16,10 @@
 
 package com.exonum.binding.cryptocurrency;
 
-import com.exonum.binding.crypto.PublicKey;
-import com.exonum.binding.hash.HashCode;
-import com.exonum.binding.messages.Transaction;
+import com.exonum.binding.common.crypto.PublicKey;
+import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.service.Service;
+import com.exonum.binding.transaction.Transaction;
 import java.util.Optional;
 
 public interface CryptocurrencyService extends Service {