+++ exonum-java-binding-cryptocurrency-demo/src/test/java/com/exonum/binding/cryptocurrency/transactions/CryptocurrencyTransactionTemplate.java
@@ -16,8 +16,8 @@
 
 package com.exonum.binding.cryptocurrency.transactions;
 
+import com.exonum.binding.common.message.Message;
 import com.exonum.binding.cryptocurrency.CryptocurrencyService;
-import com.exonum.binding.messages.Message;
 import java.nio.ByteBuffer;
 
 final class CryptocurrencyTransactionTemplate {